+++ crates/ui/src/displays/main.tsx
@@ -3,9 +3,12 @@ import "../styles/main.css";
 import { App } from "./app";
 import { EditorProvider } from "@/contexts/editor-context";
 import ReactDOM from "react-dom/client";
+import { SettingsProvider } from "@/contexts/settings-context";
 
 ReactDOM.createRoot(document.getElementById("root") as HTMLElement).render(
-  <EditorProvider>
-    <App />
-  </EditorProvider>
+  <SettingsProvider>
+    <EditorProvider>
+      <App />
+    </EditorProvider>
+  </SettingsProvider>
 );